--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -32,7 +32,7 @@ jobs:
           name: dependencies
           command: |
             python3.7 -m pip install tox
-            rustup default nightly-2019-02-28
+            rustup default nightly-2019-06-04
             rustup target add x86_64-unknown-linux-musl
       - run:
           name: build-wheels