--- a/ci/azure/install.cmd
+++ b/ci/azure/install.cmd
@@ -24,6 +24,8 @@ rustup-init.exe -y --default-toolchain nightly-2019-02-28
 set PATH=%PATH%;%USERPROFILE%\.cargo\bin
 echo "##vso[task.setvariable variable=PATH;]%PATH%;%USERPROFILE%\.cargo\bin"
 
+rustup default nightly-2019-02-28
+
 @rem Install the build and runtime dependencies of the project.
 cd python/
 pip install -r requirements.txt