--- a/ci/azure/install.sh
+++ b/ci/azure/install.sh
@@ -18,7 +18,7 @@ python -c "import numpy; print('numpy %s' % numpy.__version__)"
 python -c "import scipy; print('scipy %s' % scipy.__version__)"
 pip list
 
-curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain nightly-2019-02-28
+curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain nightly-2019-06-04
 source $HOME/.cargo/env
 
 cd python/