--- a/ci/azure/test_python.cmd
+++ b/ci/azure/test_python.cmd
@@ -5,4 +5,5 @@ call activate %VIRTUALENV%
 mkdir %TMP_FOLDER%
 cd %TMP_FOLDER%
 
-pytest --junitxml=%JUNITXML% --showlocals --durations=20 %PYTEST_ARGS% --pyargs vtext
+set RUST_BACKTRACE=1
+pytest -sv --junitxml=%JUNITXML% --showlocals --pyargs vtext