--- a/python/Cargo.toml
+++ b/python/Cargo.toml
@@ -14,11 +14,12 @@ vtext = {"path" = "../"}
 rust-stemmers = "1.1.0"
 
 [dependencies.numpy]
-version = "0.5.0"
+git = "https://github.com/kngwyu/rust-numpy.git"
+branch = "release0.6"
 features = ["python3"]
 
 [dependencies.pyo3]
-version = "0.6.0"
+version = "0.7.0"
 
 [features]
 extension-module = ["pyo3/extension-module"]