--- a/python/Cargo.toml
+++ b/python/Cargo.toml
@@ -8,18 +8,17 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-ndarray = "~0.12.0"
-sprs = "0.6.2"
+ndarray = "0.12"
+sprs = "0.6"
 vtext = {"path" = "../"}
-rust-stemmers = "1.1.0"
+rust-stemmers = "1.1"
 
 [dependencies.numpy]
-git = "https://github.com/kngwyu/rust-numpy.git"
-branch = "release0.6"
+version = "0.6"
 features = ["python3"]
 
 [dependencies.pyo3]
-version = "0.7.0"
+version = "0.7"
 
 [features]
 extension-module = ["pyo3/extension-module"]