@rth 2019-05-26
Hmm, no actually creating a `PyList` from `Vec<&str>` works but segfaults on Windows (probably due to the use of `unsafe` in Pyo3) and the fact that lifetimes are not right. Revering the change to tokenizers, unfortunately, though it should be possible to optimize this further.